+++ tests/test_function_manager.py
@@ -41,6 +41,7 @@ class TestAbilityResolution:
             mgr._enabled_tools = []
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.current_toolset_name = "none"
 
             with patch('core.chat.function_manager.toolset_manager') as mock_ts:
@@ -59,6 +60,7 @@ class TestAbilityResolution:
             mgr._enabled_tools = mgr.all_possible_tools.copy()
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.function_modules = {}
             mgr.current_toolset_name = "all"
             
@@ -87,6 +89,7 @@ class TestAbilityResolution:
             mgr._enabled_tools = []
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.current_toolset_name = "none"
             
             with patch('core.chat.function_manager.toolset_manager') as mock_ts:
@@ -113,6 +116,7 @@ class TestAbilityResolution:
             mgr._enabled_tools = []
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.current_toolset_name = "none"
             
             with patch('core.chat.function_manager.toolset_manager') as mock_ts:
@@ -139,6 +143,7 @@ class TestAbilityResolution:
             mgr._enabled_tools = []
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.current_toolset_name = "none"
 
             with patch('core.chat.function_manager.toolset_manager') as mock_ts:
@@ -189,6 +194,7 @@ class TestCustomSentinelReApply:
             ]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.current_toolset_name = "custom"
             return mgr
 
@@ -217,6 +223,7 @@ class TestCustomSentinelReApply:
             mgr._enabled_tools = []
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.current_toolset_name = "custom"
             with patch('core.chat.function_manager.toolset_manager') as mock_ts:
                 mock_ts.toolset_exists.return_value = False
@@ -240,6 +247,7 @@ class TestNewToolAutoJoinsActiveSavedToolset:
             mgr._enabled_tools = []
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._network_functions = set()
             mgr._is_local_map = {}
             mgr._function_module_map = {}
@@ -383,9 +391,11 @@ class TestExecution:
             mgr._enabled_tools = [{'function': {'name': 'test_func'}}]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.execution_map = {'test_func': mock_executor}
             mgr._is_local_map = {'test_func': True}
             mgr._function_module_map = {}
@@ -407,9 +417,11 @@ class TestExecution:
             mgr._enabled_tools = [{'function': {'name': 'allowed_func'}}]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.execution_map = {'disabled_func': MagicMock()}
             mgr._is_local_map = {}
             mgr._function_module_map = {}
@@ -437,9 +449,11 @@ class TestExecution:
             mgr._enabled_tools = [{'function': {'name': 'orphan_func'}}]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.execution_map = {}
             mgr._is_local_map = {'orphan_func': True}
             mgr._function_module_map = {}
@@ -463,9 +477,11 @@ class TestExecution:
             mgr._enabled_tools = [{'function': {'name': 'crashy_func'}}]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.execution_map = {'crashy_func': failing_executor}
             mgr._is_local_map = {'crashy_func': True}
             mgr._function_module_map = {}
@@ -497,9 +513,11 @@ class TestNetworkToolDetection:
             ]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._network_functions = {'web_search', 'web_fetch'}
 
             assert mgr.has_network_tools_enabled() is True
@@ -515,9 +533,11 @@ class TestNetworkToolDetection:
             ]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._network_functions = {'web_search', 'web_fetch'}
 
             assert mgr.has_network_tools_enabled() is False
@@ -560,6 +580,7 @@ class TestModeFiltering:
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.function_modules = {
                 'test_module': {'available_functions': ['mono_only', 'assembled_only', 'both_modes']},
                 'other_module': {'available_functions': ['no_filter']},
@@ -593,6 +614,7 @@ class TestModeFiltering:
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.function_modules = {
                 'test_module': {'available_functions': ['mono_only', 'assembled_only', 'both_modes']},
             }
@@ -616,9 +638,11 @@ class TestModeFiltering:
             ]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.function_modules = {}
 
             filtered = mgr.enabled_tools
@@ -645,9 +669,11 @@ class TestEnabledFunctionNames:
             ]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
 
             names = mgr.get_enabled_function_names()
 
@@ -661,9 +687,11 @@ class TestEnabledFunctionNames:
             mgr._enabled_tools = []
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
 
             names = mgr.get_enabled_function_names()
 
@@ -689,9 +717,11 @@ class TestAbilityInfo:
             ]
             mgr._mode_filters = {}
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr._story_engine = None
             mgr._story_engine_enabled = False
             mgr._settings_gates = {}
+            mgr._hidden_tools = set()
             mgr.function_modules = {
                 'web': {'available_functions': ['search', 'fetch']}
             }